Added test project Added RaccineLib project Added example test
