+++ .gitignore
@@ -71,3 +71,5 @@ raccine.ilk
 Raccine/*
 Raccine.zip
 build_dist.bat
+
+packages/
