+++ raccine.cpp
@@ -5,524 +5,11 @@
 // Florian Roth, Ollie Whitehouse, Branislav Djalic, John Lambert
 // with help of Hilko Bengen
 
-#include <cwchar>
-#include <Windows.h>
-#include <TlHelp32.h>
-#include <cstdio>
-#include <cstring>
-#include <clocale>
-#include <Psapi.h>
-#include <string>
-#include <array>
-#include <chrono>
-#include <ctime>
-#include <iomanip>
-#include <sstream>
-#include <strsafe.h>
+#include "source/RaccineLib/Raccine.h"
+#include "source/RaccineLib/Utils.h"
+
 #include <Shlwapi.h>
-#include <vector>
-
-
-#include "HandleWrapper.h"
-#include "YaraRuleRunner.h"
-#include "Utils.h"
-
-#pragma comment(lib,"advapi32.lib")
-#pragma comment(lib,"shlwapi.lib")
-
-// Version
-#define VERSION "1.0.4 BETA"
-
-// Log Config and Flags
-BOOL g_fLogOnly = FALSE;
-BOOL g_fShowGui = FALSE;
-#define RACCINE_REG_CONFIG  L"SOFTWARE\\Raccine"
-#define RACCINE_REG_POICY_CONFIG  L"SOFTWARE\\Policies\\Raccine"
-constexpr UINT MAX_MESSAGE = 1000;
-#define RACCINE_DEFAULT_EVENTID  1
-#define RACCINE_EVENTID_MALICIOUS_ACTIVITY  2
-
-#define RACCINE_DATA_DIRECTORY  L"%PROGRAMDATA%\\Raccine"
-#define RACCINE_PROGRAM_DIRECTORY  L"%PROGRAMFILES%\\Raccine"
-WCHAR g_wRaccineDataDirectory[MAX_PATH] = { 0 };  // ENV expanded RACCINE_DATA_DIRECTORY
-WCHAR g_wRaccineProgramDirectory[MAX_PATH] = { 0 };  // ENV expanded RACCINE_PROGRAM_DIRECTORY
-
-// YARA Matching
-WCHAR g_wYaraRulesDir[MAX_PATH] = { 0 };
-
-constexpr UINT MAX_YARA_RULE_FILES = 200;
-#define RACCINE_REG_CONFIG  L"SOFTWARE\\Raccine"
-#define RACCINE_REG_POICY_CONFIG  L"SOFTWARE\\Policies\\Raccine"
-#define RACCINE_YARA_RULES_PATH L"RulesDir"
-#define RACCINE_DEFAULT_EVENTID  1
-#define RACCINE_EVENTID_MALICIOUS_ACTIVITY  2
-
-enum class Integrity
-{
-    Error = 0, // Indicates integrity level could not be found
-    Low = 1,
-    Medium = 2,
-    High = 3,
-    System = 4,
-
-};
-
-/// <summary>
-/// Evaluate a set of yara rules on a command line
-/// </summary>
-/// <param name="lpCommandLine">The command line to test</param>
-/// <param name="outYaraOutput">if not empty, an output string containing match results is written to this parameter.</param>
-/// <returns>TRUE if at least one match result was found</returns>
-BOOL EvaluateYaraRules(LPWSTR lpCommandLine, std::wstring& outYaraOutput)
-{
-    BOOL fRetVal = FALSE;
-    WCHAR wTestFilename[MAX_PATH] = { 0 };
-    const int len = static_cast<int>(wcslen(lpCommandLine));
-    LPSTR lpAnsiCmdLine = static_cast<LPSTR>(LocalAlloc(LPTR, len + 1));
-    if (!lpAnsiCmdLine)
-    {
-        return FALSE;
-    }
-    ExpandEnvironmentStringsW(RACCINE_DATA_DIRECTORY, wTestFilename, ARRAYSIZE(wTestFilename) - 1);
-    YaraRuleRunner rule_runner(wTestFilename, g_wRaccineProgramDirectory);
-
-    int c = GetTempFileNameW(wTestFilename, L"Raccine", 0, wTestFilename);
-    if (c != 0)
-    {
-        //  Creates the new file to write to for the upper-case version.
-        HANDLE hTempFile = CreateFileW(wTestFilename, // file name 
-                                       GENERIC_WRITE,        // open for write 
-                                       0,                    // do not share 
-                                       NULL,                 // default security 
-                                       CREATE_ALWAYS,        // overwrite existing
-                                       FILE_ATTRIBUTE_NORMAL,// normal file 
-                                       NULL);                // no template 
-        if (hTempFile == INVALID_HANDLE_VALUE)
-        {
-            return FALSE;
-        }
-        DWORD dwWritten = 0;
-
-        if (WideCharToMultiByte(
-            CP_ACP,
-            0,
-            lpCommandLine,
-            len,
-            lpAnsiCmdLine,
-            len + 1,
-            NULL,
-            NULL
-        ))
-        {
-            if (!WriteFile(hTempFile, lpAnsiCmdLine, lstrlenA(lpAnsiCmdLine) + 1, &dwWritten, NULL))
-            {
-                CloseHandle(hTempFile);
-                goto cleanup;
-            }
-        }
-        CloseHandle(hTempFile);
-
-        fRetVal = rule_runner.run_yara_rules_on_file(wTestFilename, lpCommandLine, outYaraOutput);
-
-        DeleteFileW(wTestFilename);
-    }
-cleanup:
-    return fRetVal;
-}
-
-/// This function will optionally log messages to the eventlog
-void WriteEventLogEntryWithId(LPWSTR pszMessage, DWORD dwEventId)
-{
-    EventSourceHandleWrapper hEventSource = RegisterEventSourceW(NULL, L"Raccine");
-    if (!hEventSource) {
-        return;
-    }
-
-    LPCWSTR lpszStrings[2] = { NULL, NULL };
-
-    lpszStrings[0] = pszMessage;
-    lpszStrings[1] = NULL;
-
-
-    ReportEventW(hEventSource,      // Event log handle
-                 EVENTLOG_INFORMATION_TYPE,  // Event type
-                 0,                          // Event category
-                 dwEventId,                  // Event identifier
-                 NULL,                       // No security identifier
-                 1,                          // Size of lpszStrings array
-                 0,                          // No binary data
-                 lpszStrings,                // Array of strings
-                 NULL                        // No binary data
-    );
-}
-
-void WriteEventLogEntry(LPWSTR  pszMessage)
-{
-    WriteEventLogEntryWithId(pszMessage, RACCINE_DEFAULT_EVENTID);
-}
-
-// Get Parent Process ID
-DWORD getParentPid(DWORD pid)
-{
-    SnapshotHandleWrapper hSnapshot = CreateToolhelp32Snapshot(TH32CS_SNAPPROCESS, 0);
-
-    if (!hSnapshot) {
-        return 0;
-    }
-
-    PROCESSENTRY32W pe32{};
-    ZeroMemory(&pe32, sizeof(pe32));
-    pe32.dwSize = sizeof(pe32);
-    if (!Process32FirstW(hSnapshot, &pe32)) {
-        return 0;
-    }
-
-    do {
-        if (pe32.th32ProcessID == pid) {
-            return pe32.th32ParentProcessID;
-        }
-    } while (Process32NextW(hSnapshot, &pe32));
-
-    return 0;
-}
-
-// Get integrity level of process
-Integrity getIntegrityLevel(HANDLE hProcess)
-{
-
-    TokenHandleWrapper hToken = INVALID_HANDLE_VALUE;
-
-    if (!OpenProcessToken(hProcess, TOKEN_QUERY, &hToken)) {
-        return Integrity::Error;
-    }
-
-    PTOKEN_MANDATORY_LABEL pTIL;
-    DWORD dwLengthNeeded = sizeof pTIL;
-    GetTokenInformation(hToken, TokenIntegrityLevel, NULL, 0, &dwLengthNeeded);
-    pTIL = static_cast<PTOKEN_MANDATORY_LABEL>(LocalAlloc(0, dwLengthNeeded));
-    if (!pTIL) {
-        return Integrity::Error;
-    }
-
-    if (GetTokenInformation(hToken, TokenIntegrityLevel,
-                            pTIL, dwLengthNeeded, &dwLengthNeeded)) {
-        const DWORD dwIntegrityLevel = *GetSidSubAuthority(pTIL->Label.Sid,
-                                                           static_cast<DWORD>(static_cast<UCHAR>(*GetSidSubAuthorityCount(pTIL->Label.Sid) - 1)));
-
-        LocalFree(pTIL);
-
-        if (dwIntegrityLevel == SECURITY_MANDATORY_LOW_RID) {
-            return Integrity::Low;
-        }
-
-        if (dwIntegrityLevel >= SECURITY_MANDATORY_MEDIUM_RID &&
-            dwIntegrityLevel < SECURITY_MANDATORY_HIGH_RID) {
-            return Integrity::Medium;
-        }
-
-        if (dwIntegrityLevel >= SECURITY_MANDATORY_HIGH_RID &&
-            dwIntegrityLevel < SECURITY_MANDATORY_SYSTEM_RID) {
-            return Integrity::High;
-        }
-
-        if (dwIntegrityLevel >= SECURITY_MANDATORY_SYSTEM_RID) {
-            return Integrity::System;
-        }
-
-        return Integrity::Error;
-    }
-
-    LocalFree(pTIL);
-    return Integrity::Error;
-}
-
-// Get the image name of the process
-std::wstring getImageName(DWORD pid)
-{
-    PROCESSENTRY32W pe32{};
-    SnapshotHandleWrapper hSnapshot = CreateToolhelp32Snapshot(TH32CS_SNAPPROCESS, 0);
-
-    if (!hSnapshot) {
-        return L"(unavailable)";
-    }
-
-    ZeroMemory(&pe32, sizeof(pe32));
-    pe32.dwSize = sizeof(pe32);
-
-    if (!Process32FirstW(hSnapshot, &pe32)) {
-        return L"(unavailable)";
-    }
-
-    do {
-        if (pe32.th32ProcessID == pid) {
-            return std::wstring(static_cast<wchar_t*>(pe32.szExeFile));
-        }
-    } while (Process32NextW(hSnapshot, &pe32));
-
-    return L"(unavailable)";
-}
-
-// Helper for isAllowListed, checks if a specific process is allowed
-bool isProcessAllowed(const PROCESSENTRY32W& pe32)
-{
-    ProcessHandleWrapper hProcess = OpenProcess(PROCESS_QUERY_INFORMATION, FALSE, pe32.th32ProcessID);
-    if (!hProcess) {
-        return false;
-    }
-
-    const std::array<std::wstring, 3> allow_list{ L"wininit.exe", L"winlogon.exe", L"explorer.exe" };
-    for (const std::wstring& allowed_name : allow_list) {
-        if (_wcsicmp(static_cast<const wchar_t*>(pe32.szExeFile), allowed_name.c_str()) != 0) {
-            continue;
-        }
-
-        wchar_t filePath[MAX_PATH] = { 0 };
-        if (GetModuleFileNameEx(hProcess, NULL, filePath, MAX_PATH)) {
-            // Are they in the Windows directory?
-            const std::wstring system32_path = L"C:\\Windows\\System32\\";
-            if (_wcsnicmp(filePath, system32_path.c_str(), system32_path.length()) == 0) {
-                // Is the process running as SYSTEM
-                return getIntegrityLevel(hProcess) == Integrity::System;
-            }
-
-            // Are you explorer running in the Windows dir
-            const std::wstring explorer_path = L"C:\\Windows\\Explorer.exe";
-            if (_wcsnicmp(filePath, explorer_path.c_str(), explorer_path.length()) == 0) {
-                // Is the process running as MEDIUM (which Explorer does)
-                return getIntegrityLevel(hProcess) == Integrity::Medium;
-            }
-        }
-    }
-
-    return false;
-}
-
-// Check if process is in allowed list
-bool isAllowListed(DWORD pid)
-{
-    SnapshotHandleWrapper hSnapshot = CreateToolhelp32Snapshot(TH32CS_SNAPPROCESS, 0);
-
-    if (!hSnapshot) {
-        return false;
-    }
-
-    PROCESSENTRY32W pe32{};
-    pe32.dwSize = sizeof pe32;
-
-    if (!Process32FirstW(hSnapshot, &pe32)) {
-        return false;
-    }
-
-    do {
-        if (pe32.th32ProcessID != pid) {
-            continue;
-        }
-
-        return isProcessAllowed(pe32);
-    } while (Process32NextW(hSnapshot, &pe32));
-
-    return false;
-}
-
-// Kill a process
-BOOL killProcess(DWORD dwProcessId, UINT uExitCode) {
-    constexpr DWORD dwDesiredAccess = PROCESS_TERMINATE;
-    constexpr BOOL  bInheritHandle = FALSE;
-    ProcessHandleWrapper hProcess = OpenProcess(dwDesiredAccess, bInheritHandle, dwProcessId);
-    if (!hProcess) {
-        return FALSE;
-    }
-
-    return TerminateProcess(hProcess, uExitCode);
-}
-
-// Get timestamp
-std::string getTimeStamp() {
-    struct tm buf {};
-    auto time = std::chrono::system_clock::to_time_t(std::chrono::system_clock::now() - std::chrono::hours(24));
-    localtime_s(&buf, &time);
-    std::stringstream ss;
-    ss << std::put_time(&buf, "%F %T");
-    auto timestamp = ss.str();
-    return timestamp;
-}
-
-// Format a log lines
-std::wstring logFormat(const std::wstring& cmdLine, const std::wstring& comment = L"done") {
-    const std::string timeString = getTimeStamp();
-    const std::wstring timeStringW(timeString.cbegin(), timeString.cend());
-    std::wstring logLine = timeStringW + L" DETECTED_CMD: '" + cmdLine + L" COMMENT: " + comment + L"\n";
-    return logLine;
-}
-
-std::wstring logFormatLine(const std::wstring& line = L"") {
-    const std::string timeString = getTimeStamp();
-    const std::wstring timeStringW(timeString.cbegin(), timeString.cend());
-    std::wstring logLine = timeStringW + L" " + line + L"\n";
-    return logLine;
-}
-
-// Format the activity log lines
-std::wstring logFormatAction(DWORD pid, const std::wstring& imageName, const std::wstring& cmdLine, const std::wstring& comment = L"done") {
-    const std::string timeString = getTimeStamp();
-    const std::wstring timeStringW(timeString.cbegin(), timeString.cend());
-    std::wstring logLine = timeStringW + L" DETECTED_CMD: '" + cmdLine + L"' IMAGE: '" + imageName + L"' PID: " + std::to_wstring(pid) + L" ACTION: " + comment + L"\n";
-    return logLine;
-}
-
-// Log to file
-void logSend(const std::wstring& logStr) {
-    static FILE* logFile = nullptr;
-    if (logFile == nullptr) {
-        errno_t err = _wfopen_s(&logFile, L"C:\\ProgramData\\Raccine\\Raccine_log.txt", L"at");
-
-        if (err != 0) {
-            err = _wfopen_s(&logFile, L"C:\\ProgramData\\Raccine\\Raccine_log.txt", L"wt");
-        }
-
-        if (err != 0) {
-            wprintf(L"\nCan not open C:\\ProgramData\\Raccine\\Raccine_log.txt for writing.\n");
-            return;   // bail out if we can't log
-        }
-    }
-    //transform(logStr.begin(), logStr.end(), logStr.begin(), ::tolower);
-    if (logFile != nullptr)
-    {
-        fwprintf(logFile, L"%s", logStr.c_str());
-        fflush(logFile);
-        fclose(logFile);
-        logFile = nullptr;
-    }
-}
-
-//
-//  Query for config in HKLM and HKLM\Software\Policies override by GPO
-//
-void InitializeSettings()
-{
-    // Registry Settings
-    // Query for logging level. A value of 1 or more indicates to log key events to the event log
-    // Query for logging only mode. A value of 1 or more indicates to suppress process kills
-
-    ExpandEnvironmentStrings(RACCINE_DATA_DIRECTORY, g_wRaccineDataDirectory, ARRAYSIZE(g_wRaccineDataDirectory) - 1);
-    ExpandEnvironmentStrings(RACCINE_PROGRAM_DIRECTORY, g_wRaccineProgramDirectory, ARRAYSIZE(g_wRaccineProgramDirectory) - 1);
-
-    StringCchCopy(g_wYaraRulesDir, ARRAYSIZE(g_wYaraRulesDir), g_wRaccineDataDirectory);
-
-    const wchar_t* LoggingKeys[] = { RACCINE_REG_CONFIG , RACCINE_REG_POICY_CONFIG };
-
-    HKEY hKey = NULL;
-    for (int i = 0; i < ARRAYSIZE(LoggingKeys); i++)
-    {
-        if (ERROR_SUCCESS == RegOpenKeyEx(HKEY_LOCAL_MACHINE, LoggingKeys[i], 0, KEY_READ, &hKey))
-        {
-            // Log Only
-            DWORD dwLoggingOnly = 0;
-            DWORD cbDataLO = sizeof(dwLoggingOnly);
-            if (ERROR_SUCCESS == RegQueryValueExW(hKey, L"LogOnly", NULL, NULL, (LPBYTE)&dwLoggingOnly, &cbDataLO))
-            {
-                if (dwLoggingOnly > 0)
-                {
-                    g_fLogOnly = TRUE;
-                }
-            }
-            // Show Gui
-            DWORD dwShowGui = 0;
-            DWORD cbDataGUI = sizeof(dwShowGui);
-            if (ERROR_SUCCESS == RegQueryValueExW(hKey, L"ShowGui", NULL, NULL, (LPBYTE)&dwShowGui, &cbDataGUI))
-            {
-                if (dwShowGui > 0)
-                {
-                    g_fShowGui = TRUE;
-                }
-            }
-            // Yara rules dir
-            DWORD cbData = sizeof(g_wYaraRulesDir);
-            if (ERROR_SUCCESS == RegQueryValueExW(hKey, RACCINE_YARA_RULES_PATH, NULL, NULL, (LPBYTE)g_wYaraRulesDir, &cbData))
-            {
-                ;
-            }
-            RegCloseKey(hKey);
-        }
-    }
-}
-
-void createChildProcessWithDebugger(std::wstring command_line)
-{
-    STARTUPINFO info = { sizeof(info) };
-    PROCESS_INFORMATION processInfo{};
-
-    constexpr LPCWSTR NO_APPLICATION_NAME = nullptr;
-    constexpr LPSECURITY_ATTRIBUTES DEFAULT_SECURITY_ATTRIBUTES = nullptr;
-    constexpr BOOL INHERIT_HANDLES = TRUE;
-    constexpr LPVOID USE_CALLER_ENVIRONMENT = nullptr;
-    constexpr LPCWSTR USE_CALLER_WORKING_DIRECTORY = nullptr;
-    const BOOL res = CreateProcessW(NO_APPLICATION_NAME,
-                                    static_cast<LPWSTR>(command_line.data()),
-                                    DEFAULT_SECURITY_ATTRIBUTES,
-                                    DEFAULT_SECURITY_ATTRIBUTES,
-                                    INHERIT_HANDLES,
-                                    DEBUG_PROCESS | DEBUG_ONLY_THIS_PROCESS,
-                                    USE_CALLER_ENVIRONMENT,
-                                    USE_CALLER_WORKING_DIRECTORY,
-                                    &info,
-                                    &processInfo);
-    if (res == 0) {
-        return;
-    }
-
-    DebugActiveProcessStop(processInfo.dwProcessId);
-    WaitForSingleObject(processInfo.hProcess, INFINITE);
-    CloseHandle(processInfo.hProcess);
-    CloseHandle(processInfo.hThread);
-}
-
-// Find all parent processes and kill them
-void find_and_kill_processes(const std::wstring& sCommandLine, std::wstring& sListLogs)
-{
-    std::vector<DWORD> pids;
-    // Collect PIDs to kill
-    DWORD pid = GetCurrentProcessId();
-
-    while (true) {
-        pid = getParentPid(pid);
-        if (pid == 0) {
-            break;
-        }
-
-        std::wstring imageName = getImageName(pid);
-
-        if (!isAllowListed(pid)) {
-            wprintf(L"\nCollecting IMAGE %s with PID %d for a kill\n", imageName.c_str(), pid);
-            pids.push_back(pid);
-        }
-        else {
-            wprintf(L"\nProcess IMAGE %s with PID %d is on allowlist\n", imageName.c_str(), pid);
-            sListLogs.append(logFormatAction(pid, imageName, sCommandLine, L"Whitelisted"));
-        }
-    }
-
-    // Loop over collected PIDs and try to kill the processes
-    for (DWORD process_id : pids) {
-        std::wstring imageName = getImageName(process_id);
-        // If no simulation flag is set
-        if (!g_fLogOnly) {
-            // Kill
-            wprintf(L"Kill process IMAGE %s with PID %d\n", imageName.c_str(), process_id);
-            killProcess(process_id, 1);
-            sListLogs.append(logFormatAction(process_id, imageName, sCommandLine, L"Terminated"));
-        }
-        else {
-            // Simulated kill
-            wprintf(L"Simulated Kill IMAGE %s with PID %d\n", imageName.c_str(), process_id);
-            sListLogs.append(logFormatAction(process_id, imageName, sCommandLine, L"Terminated (Simulated)"));
-        }
-    }
-
-    // Finish message
-    printf("\nRaccine v%s finished\n", VERSION);
-    Sleep(5000);
-}
+#include <strsafe.h>
 
 int wmain(int argc, WCHAR* argv[])
 {
@@ -699,7 +186,7 @@ int wmain(int argc, WCHAR* argv[])
         LPCWSTR lpMessage = sCommandLine.c_str();
         if (!g_fLogOnly) {
             // Eventlog
-            StringCchPrintf(wMessage, ARRAYSIZE(wMessage), L"Raccine detected malicious activity:\n%s\n", lpMessage);
+            StringCchPrintfW(wMessage, ARRAYSIZE(wMessage), L"Raccine detected malicious activity:\n%s\n", lpMessage);
             // Log to the text log file
             sListLogs.append(logFormat(sCommandLine, L"Raccine detected malicious activity"));
         }
@@ -715,9 +202,9 @@ int wmain(int argc, WCHAR* argv[])
 
         // YARA Matches Detected
         if (fYaraRuleMatched && !szYaraOutput.empty()) {
-                StringCchPrintfW(wMessage, ARRAYSIZE(wMessage), L"\r\nYara matches:\r\n%s", szYaraOutput.c_str());
-                WriteEventLogEntryWithId(static_cast<LPWSTR>(wMessage), RACCINE_EVENTID_MALICIOUS_ACTIVITY);
-                sListLogs.append(logFormatLine(szYaraOutput));
+            StringCchPrintfW(wMessage, ARRAYSIZE(wMessage), L"\r\nYara matches:\r\n%s", szYaraOutput.c_str());
+            WriteEventLogEntryWithId(static_cast<LPWSTR>(wMessage), RACCINE_EVENTID_MALICIOUS_ACTIVITY);
+            sListLogs.append(logFormatLine(szYaraOutput));
         }
 
         // signal Event for UI to know an alert happened.  If no UI is running, this has no effect.
