+++ source/RaccineLib/Utils.cpp
