+++ source/RaccineLib/YaraRuleRunner.cpp
@@ -1,6 +1,5 @@
 ﻿#include "YaraRuleRunner.h"
 
-
 #include <fstream>
 #include <strsafe.h>
 #include <Windows.h>
