+++ source/RaccineLib/YaraRuleRunner.h
