+++ tests/Raccine-Test/pch.cpp
@@ -1,0 +1,6 @@
+//
+// pch.cpp
+// Include the standard header and generate the precompiled header.
+//
+
+#include "pch.h"
